ai-job-agent/requirements.txt
streamlit

# Core
python-dotenv
requests
os

pdfplumber

scikit-learn
numpy

sentence-transformers
torch


langchain
langchain-google-genai